made hexshapes
to keep the code for hexagons more separated from the triangle code
